Scroll down to frame/sample section on back
This turned out to be a larger pain in the butt than I thought. I can't
use raw anchor links because those links in the DOM don't exist on page
load - the DB is still fetching and the loading indicator is shown. This
means that all SSR pages/initial loads have no links, so the hash scroll
in the browser doesn't work.

To remediate this, we attach refs to those elements and attach callbacks
to them so we know when they load. When the page loads those anchors, we
check the hash in the location then and decide if we should scroll the
user to that section.